requirements.txt
argparse
PyYAML
Jinja2
apache-libcloud
virtualenv
certifi
